meta model generator
The goal is to generate a typed API for the model from
some language definition.
Something similar is already implemented for MPS in
https://github.com/modelix/api-gen
The new implementation is MPS independent.
It's a lightweight generator implemented in Kotlin that takes
JSON/YAML as an input and generates Kotlin classes.
On top of these classes other meta DSLs (e.g. for the editor)
can be implemented as internal Kotlin DSLs.

Moved from the modelix main repository to here.